+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `[project]` section of `pyproject.toml`:
     - normalize `name`
     - accept `keyword` key
+    - accept `optional-dependencies` key
 
 ## [v0.42.0] - 2023-05-01